OMG Super css h4X0r5
When moving the first spanned element in a row, using
jquery.ui.draggable or jquery.ui.sortable, the placeholder element
would be styled as the second span in the row. Thus having an incorrect
margin-left and killing the layout. This rule zeros the left margin on
the second element whenever the first element is positioned absolutely.