No more explosions!
This is actually an elegant solution.

FIrst off I dont know why the listenTo syntax was causing jQuery to
explode, but this change fixes it.

Also, what is elegant, is that I no longer rely on the open method to
create the event binding. Instead Backbone itself creates the event
binding as it should have been all along!